requirements.txt
Django==4.1.2
Pillow==9.2.0
django-bootstrap4==22.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
django-crispy-forms==1.14.0